Skip config.json

+++ agixt/Hub.py
@@ -82,6 +82,8 @@ def import_agixt_hub():
             if os.path.isdir(src_file):
                 if os.path.exists(dest_file):
                     for item in os.listdir(dest_file):
+                        if item == "config.json":
+                            continue
                         dest_item = os.path.join(dest_file, item)
                         if os.path.isfile(dest_item):
                             os.remove(dest_item)
@@ -89,6 +91,8 @@ def import_agixt_hub():
                     os.makedirs(dest_file, exist_ok=True)
 
                 for item in os.listdir(src_file):
+                    if item == "config.json":
+                        continue
                     src_item = os.path.join(src_file, item)
                     dest_item = os.path.join(dest_file, item)
                     if os.path.isdir(src_item):
@@ -98,6 +102,8 @@ def import_agixt_hub():
                     else:
                         shutil.copy2(src_item, dest_item)
             else:
+                if src_file.split("/")[-1] == "config.json":
+                    continue
                 if os.path.exists(dest_file):
                     os.remove(dest_file)
                 shutil.move(src_file, dest_file)
